+++ src/senaite/patient/content/patient.py
@@ -91,6 +91,28 @@ class IAdditionalEmailSchema(Interface):
     )
 
 
+class IRaceSchema(Interface):
+    """Schema definition for patient race
+    """
+    race = schema.Choice(
+        title=_("Race"),
+        description=_(u""),
+        source="senaite.patient.vocabularies.races",
+        required=False,
+    )
+
+
+class IEthnicitySchema(Interface):
+    """Schema definition for patient ethnicity
+    """
+    ethnicity = schema.Choice(
+        title=_("Ethnicity"),
+        description=_(u""),
+        source="senaite.patient.vocabularies.ethnicities",
+        required=False,
+    )
+
+
 class IPatientSchema(model.Schema):
     """Patient Content
     """
@@ -106,6 +128,12 @@ class IPatientSchema(model.Schema):
         title=u"Description",
         required=False
     )
+
+    # race/ethnicity fieldset
+    fieldset(
+        "race_and_ethnicity",
+        label=u"Race and Ethnicity",
+        fields=["races", "ethnicities"])
 
     # contact fieldset
     fieldset(
@@ -192,6 +220,42 @@ class IPatientSchema(model.Schema):
         source="senaite.patient.vocabularies.gender",
         default="",
         required=True,
+    )
+
+    directives.widget(
+        "races",
+        DataGridWidgetFactory,
+        allow_reorder=True,
+        auto_append=True)
+    races = DataGridField(
+        title=_(u"label_patient_races", default=u"Races"),
+        description=_(
+            "description_patient_races",
+            default=u"General race category reported by the patient "
+            u"- subject may have more than one"
+        ),
+        value_type=DataGridRow(schema=IRaceSchema),
+        required=False,
+        missing_value=[],
+        default=[],
+    )
+
+    directives.widget(
+        "ethnicities",
+        DataGridWidgetFactory,
+        allow_reorder=True,
+        auto_append=True)
+    ethnicities = DataGridField(
+        title=_(u"label_patient_ethnicities", default=u"Ethnicities"),
+        description=_(
+            "description_patient_ethnicities",
+            default=u"General ethnicity category reported by the patient "
+            u"- subject may have more than one"
+        ),
+        value_type=DataGridRow(schema=IEthnicitySchema),
+        required=False,
+        missing_value=[],
+        default=[],
     )
 
     # Contact
@@ -446,7 +510,7 @@ class Patient(Container):
 
     @security.protected(permissions.View)
     def getIdentifiers(self):
-        """Returns the birthday with the field accessor
+        """Returns the identifiers with the field accessor
         """
         accessor = self.accessor("identifiers")
         return accessor(self)
@@ -465,9 +529,37 @@ class Patient(Container):
 
     @security.protected(permissions.ModifyPortalContent)
     def setIdentifiers(self, value):
-        """Set birthdate by the field accessor
+        """Set identifiers by the field accessor
         """
         mutator = self.mutator("identifiers")
+        return mutator(self, value)
+
+    @security.protected(permissions.View)
+    def getRaces(self):
+        """Returns the patient races
+        """
+        accessor = self.accessor("races")
+        return accessor(self)
+
+    @security.protected(permissions.ModifyPortalContent)
+    def setRaces(self, value):
+        """Set the patient races
+        """
+        mutator = self.mutator("races")
+        return mutator(self, value)
+
+    @security.protected(permissions.View)
+    def getEthnicities(self):
+        """Returns the patient ethnicities
+        """
+        accessor = self.accessor("ethnicities")
+        return accessor(self)
+
+    @security.protected(permissions.ModifyPortalContent)
+    def setEthnicities(self, value):
+        """Set the patient ethnicities
+        """
+        mutator = self.mutator("ethnicities")
         return mutator(self, value)
 
     @security.protected(permissions.View)