+++ docs/changelog.rst
@@ -4,6 +4,7 @@ Changelog
 1.3.0 (unreleased)
 ------------------
 
+- #51 Add Race and Ethnicity Categories for Patients
 - #50 Allow additional email addresses for patients
 - #47 Display "Estimated" next to date of birth when estimated from age
 - #49 Group configuration settings